+++ app/components/LoginPage.tsx
@@ -10,7 +10,8 @@ import { parse } from 'query-string';
 import i18n from 'i18next';
 import { Button, Input, Message, Form, FullScreenSegment, Logo } from './basic';
 import SplashLoadingScreen from '../utils/SplashLoadingScreen';
-import LogoLabel from './LogoLabel';
+import LogoLabel from './banner/LogoLabel';
+import LargeLogo from './banner/LargeLogo';
 
 export default class LoginPage extends Component {
     constructor(props, context) {
@@ -69,15 +70,7 @@ export default class LoginPage extends Component {
         return (
             <FullScreenSegment>
                 <div className={`loginContainer ${isHeaderTextPresent ? 'loginContainerExtended' : ''}`}>
-                    <Logo
-                        style={{
-                            textAlign: 'center',
-                            margin: '0 auto',
-                            display: 'block',
-                            width: 100,
-                            height: 100
-                        }}
-                    />
+                    <LargeLogo />
                     {isHeaderTextPresent && (
                         <div style={{ textAlign: 'center', marginBottom: 30 }}>
                             {loginPageHeader && <LogoLabel color={loginPageHeaderColor} content={loginPageHeader} />}
