+++ app/components/banner/LargeLogo.tsx
@@ -1,0 +1,18 @@
+import React, { FunctionComponent } from 'react';
+import { Logo } from '../basic';
+
+const LargeLogo: FunctionComponent = () => {
+    return (
+        <Logo
+            style={{
+                textAlign: 'center',
+                margin: '0 auto',
+                display: 'block',
+                width: 100,
+                height: 100
+            }}
+        />
+    );
+};
+
+export default LargeLogo;
