+++ app/components/banner/LogoLabel.tsx
